GQL-11 feat(auth): add loading for main page
