+++ src/App.tsx
@@ -1,13 +1,16 @@
 import { Header } from './components/Header/Header';
 import { Footer } from './components/Footer/Footer';
 import { AppRoutes } from './routes/AppRoutes';
+import { useLoadUser } from './hooks/useLoadUser';
+import { CircularProgress } from '@mui/material';
 
-export const App = () => (
-  <>
-    <Header />
-    <main>
-      <AppRoutes />
-    </main>
-    <Footer />
-  </>
-);
+export const App = () => {
+  const loading = useLoadUser();
+  return (
+    <>
+      <Header />
+      <main>{(loading && <CircularProgress />) || <AppRoutes />}</main>
+      <Footer />
+    </>
+  );
+};
