+++ src/routes/AppRoutes.tsx
@@ -1,27 +1,13 @@
 import { Navigate, Route, Routes } from 'react-router';
-import { useEffect } from 'react';
 
 import { Main } from '../pages/Main/Main';
 import { NotFound } from '../pages/NotFound/NotFound';
 import { useAuth } from '../hooks/useAuth';
-import { useAppDispatch } from '../hooks/redux-hooks';
-import { auth } from '../firebase/firebase-config';
-import { onAuthStateChanged } from 'firebase/auth';
-import { setUser } from '../redux/slices/userSlice';
 import { Login } from '../pages/Login/Login';
 import { Register } from '../pages/Register/Register';
 
 export const AppRoutes = () => {
   const { isAuth } = useAuth();
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    onAuthStateChanged(auth, (user) => {
-      if (user) {
-        dispatch(setUser({ email: user.email, id: user.uid, token: user.refreshToken }));
-      }
-    });
-  }, [dispatch]);
 
   return (
     <Routes>
